fix login with phone

+++ Models/DTO/Booking/BookingRevenueResponseDto.cs
@@ -2,11 +2,11 @@ namespace GraduationThesis_CarServices.Models.DTO.Booking
 {
     public class BookingRevenueResponseDto
     {
-        public double AmountEarned {get; set;}
-        public double ServiceEarned {get; set;}
-        public double ProductEarned {get; set;}
-        public double SumPaid {get; set;}
-        public double SumUnPaid {get; set;}
+        public decimal AmountEarned {get; set;}
+        public decimal ServiceEarned {get; set;}
+        public decimal ProductEarned {get; set;}
+        public decimal SumPaid {get; set;}
+        public decimal SumUnPaid {get; set;}
         public int CountPaid {get; set;}
         public int CountUnpaid {get; set;}
     }
